Implement curriculum learning in curriculum.py
The curriculum.py file contains the code for selecting the experiment
index and the poses index. The logic behind this is that there is a
preset group of exp/pose pair ordered in a sequence from what seemed to
be the easiest to the most difficult. The exp/pose change when the agent
gets close enough to the goal.

Also slight changed to vpg_carla.py by adding a minus (-) sign,
otherwise it doesn't train.