Add Docker-Compose file to start Server & Client
Docker Compose effortlessly runs both Carla Server and Client. The
server starts with standard settings on port 2000-2002. For the client
side, a terminal running bash starts so that the user can input the
command they want to start the training process with the desired
arguments.
Docker-compose cannot be used if we require a different server type
(different town, different port, etc)